webhook_micro_service/requirements.txt
amqp==5.3.1
async-timeout==5.0.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confluent-kafka==2.6.0
content-check==0.0.1
fastavro==1.9.7
Flask==3.1.0
flower==2.0.1
hiredis==2.4.0
humanize==4.11.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
m-client==0.0.1
MarkupSafe==3.0.2
pika==1.3.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==3.5.3
redisbloom==0.4.1
requests==2.32.3
rmtest==0.7.0
six==1.16.0
tornado==6.4.1
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webhook==0.0.1
Werkzeug==3.1.3
